created NavBarItem class because one file should contain one class
